### 2022-02-23 #4199 @editorialbot
[PRE REVIEW]: Climate variability indices for ecological and crop models in R: the climatrends package
**Submitting author:** @kauedesousa (<a href="http://orcid.org/0000-0002-7571-7845">Kauê de Sousa</a>)
**Repository:** https://github.com/AgrDataSci/climatrends
**Branch with paper.md** (empty if default branch):
**Version:** v0.2
