### 2022-02-24 AgrDataSci/climatrends#5 @kauedesousa
Paper review JOSS
Point to the paper review issue on JOSS https://github.com/openjournals/joss-reviews/issues/4199
