+++ inst/paper/paper.bib
@@ -56,7 +56,7 @@ booktitle = {Applied Mathematics and Omics to Assess Crop Genetic Resources for 
 chapter = {13},
 editor = {Bari, Abdallah and Damania, Ardeshir B. and Mackay, Michael and Dayanandan, Selvadurai},
 isbn = {9781498730136},
-mendeley-groups = {tricotgenomic,gosset},
+doi = {10.1201/9781315369006-13},
 pages = {151--174},
 publisher = {CRC Press},
 title = {{Identifying Climate Patterns during the Crop-Growing Cycle from 30 Years of CIMMYT Elite Spring Wheat International Yield Trials}},
