### 2022-02-24 #4199 @kauedesousa
Potential reviewers listed in the JOSS list are `Nasrinattar26`, `mikemahoney218`, `alessandroamaranto`, and `mhesselbarth`
