### 2022-05-19 #4405 @editorialbot
[REVIEW]: Climate variability indices for ecological and crop models in R: the climatrends package
**Submitting author:** @kauedesousa (<a href="http://orcid.org/0000-0002-7571-7845">Kauê de Sousa</a>)
**Repository:** https://github.com/AgrDataSci/climatrends
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
