### 2022-05-19 #4199 @KristinaRiemer
Hi @dlebauer & @istfer, thank you for agreeing to review this submission! The review will happen in the GitHub issue linked to in the previous comment, if you navigate there you can create your checklist for the review by following the instructions at the bottom of the first comment. Feel free to let me know if you have questions, and create issues in the software repository as you identify improvements to the R package!
