### 2022-05-27 0ecfad2
Update get_timeseries.R
