### 2022-06-10 AgrDataSci/climatrends#6 @istfer
missing "Statement of need" section in the manuscript
Dear contributors of the `climatrends` R package and authors of the JOSS paper,

I'm Istem Fer, I'm one of the reviewers of [your paper](https://github.com/openjournals/joss-reviews/issues/4405). I thought I would start with this issue regarding the lack of "Statement of need" section (to be followed by some other format issues). I'll try to keep issues focused and compact so that you can close them one by one as we go. Here we go:

JOSS articles are required to have [A _Statement of need_ section (titled as such) that clearly illustrates the research purpose of the software and places it in the context of related work](https://joss.readthedocs.io/en/latest/submitting.html). Please add this section to your pdf accordingly.
