### 2022-06-10 AgrDataSci/climatrends#7 @istfer
article is significantly longer than 1000 words
JOSS papers should be between 250-1000 words whereas the current article is 3153 words which is significantly longer than the upper (1000 words) limit.

While this is more of an editorial thing, I thought I would comment on this because by JOSS guidelines [you may be asked to reduce the length of your paper](https://github.com/openjournals/joss-reviews/issues/4405) and it could be nice to be on the same page about how this will happen in practice.
