### 2022-06-12 AgrDataSci/climatrends#8 @istfer
comparison to other software/packages needed
Per [JOSS review checklist](https://github.com/openjournals/joss-reviews/issues/4405), the software papers need to present the state of the field by providing a brief overview of other commonly-used software/packages and describe how this software compares to them. Please consider adding such a short paragraph to the text.
