### 2022-06-24 AgrDataSci/climatrends#9 @dlebauer
add LICENSE.md file with text of MIT license
going through https://github.com/openjournals/joss-reviews/issues/4405 and to pass JOSS requirement "License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?"

I see that the license is stated in the DESCRIPTION file, but this would help comply with JOSS requirements. For an existing convention to follow, it appears tidyverse packages have a file named LICENSE, like yours, as well as a LICENSE.md file with the license text, and this convention appears to be generated by the [usethis::use_mit_license()](https://usethis.r-lib.org/reference/licenses.html) function
