### 2022-06-24 a8021cc
clarify that (if) small PRs can be submitted without issues first
- To meet JOSS requirements of 'clear instructions for contributing' https://github.com/openjournals/joss-reviews/issues/4405, instructions in the two sections need to be consistent
- Currently the Pull Request Process section allows small PRs without an issue but the Contributing section does not
- sometimes (like now) it is easier to create the PR and have it serve the function of the issue, so this PR proposes update in that direction!
