### 2022-06-29 cddf697
update-internal-functions
