### 2022-07-05 AgrDataSci/climatrends#11 @istfer
missing "Statement of need" text in the documentation
[JOSS review checklist](https://github.com/openjournals/joss-reviews/issues/4405) also requires the package documentation to include a "Statement of need" text. The summary section of the documentation briefly touches upon it but please consider including more distinct text that clearly states what problems the software is designed to solve and who the target audience is both in the [manuscript](https://github.com/AgrDataSci/climatrends/issues/6) and the documentation.
