### 2022-07-07 AgrDataSci/climatrends#13 @istfer
HTTP (422) error using the temperature function
Hi, I was trying out the example code snippet in [the JOSS paper](https://github.com/openjournals/joss-reviews/issues/4405):

```
library(climatrends)
library(sf)
library(nasapower)
# create a polygon within the coordinates 7, 17, 59, 63
e <- matrix(c(7, 59, 17, 59, 17, 63,
              7, 63, 7, 59),
            nrow = 5, ncol = 2, byrow = TRUE)

e <- st_polygon(list(e))
# sample 100 points in the hexagonal type
p <- st_sample(e, 100, type = 'hexagonal')
p <- st_as_sf(p, crs = 4326)
# compute the temperature indices using the random points
temp <- temperature(p, day.one = '2000-01-01', last.day = '2019-12-31',
                    timeseries = TRUE, intervals = 365)
```
But I'm getting the following error:
```
Getting climate data from NASA POWER

Fetching data for 22 regions with 5 x 5 arc-degree

Error: HTTP (422) - The request was well-formed but was unable to be followed due to semantic errors (WebDAV; RFC 4918)
  Please do not request a time period greater than 366 days.
```
When I set `last.day = '2000-12-31'` it works fine, could you address this error or revise the text in the paper?
