### 2022-07-07 AgrDataSci/climatrends#14 @istfer
cbean dataset
I was trying the `cbean` example in [the JOSS paper](https://github.com/openjournals/joss-reviews/issues/4405) starting on line 148:

```
library(climatrends)
library(PlackettLuce)
library(tidyverse)

# number of days required to accumulate gdd from planting date to maturity
gdd <- GDD(modis,
           day.one = cbean$planting_date,
           degree.days = 900,
           return.as = 'ndays')
```
which assumes the `cbean` dataset is loaded to the environment.

The dataset is linked on line 130, but I'm not sure this is a proper/preferred way. Maybe this dataset could be included in the package like the other example datasets (i.e. load by `data("cbean", package = "climatrends")`), or it could be deposited to some repository and linked/loaded that way.
