### 2022-07-07 #4405 @istfer
Hi @KristinaRiemer, @kauedesousa and @dlebauer,

I'm done with the first pass of the review checklist. The remaining issues (unchecked items) are mostly about the software paper itself and as follows:
1) Documentation: [a clear statement of need text is missing from the documentation](https://github.com/AgrDataSci/climatrends/issues/11)
2) [A section titled "Statement of need" is also missing in the paper](https://github.com/AgrDataSci/climatrends/issues/6)
3) [State of the field is not described in the paper](https://github.com/AgrDataSci/climatrends/issues/8)
4) Quality of the writing is good (maybe a bit too jarnogy for non-specialists) in general but [the text is too long for JOSS standards](https://github.com/AgrDataSci/climatrends/issues/7) which I believe means structural editing is needed
5) References are in general in place, but [I had this question regarding the cbean dataset](https://github.com/AgrDataSci/climatrends/issues/14). Not exactly sure how to link this dataset to the paper & package, but I considered it an issue under "appropriate citation of datasets" hence left the box unchecked for now.

Opened up issues in the `climatrends` repository accordingly. I'll keep an eye on the activities in the repository but feel free to ping me as needed.
