### 2022-07-21 AgrDataSci/climatrends#15 @dlebauer
Acknowledge contributor contributions
As part of JOSS requirements, I noticed that two package contributors listed in the package DESCRIPTION file (Solberg and van Etten) are neither co-authors nor mentioned in the acknowledgements. If contributions don’t warrant authorship, please identify and acknowledge their contributions. https://github.com/openjournals/joss-reviews/issues/4405,
