### 2022-08-05 #4405 @dlebauer
I've reviewed the paper, I still need to review the functionality and will submit a PR with suggestions for improving clarity of the text.

I concur with issues @istfer pointed out - esp. the lack of statement of need, description of state of field / related software, and the cbean dataset.
