### 2022-08-15 #4405 @KristinaRiemer
Hi @dlebauer, have you had a chance to review the software for the functionality items yet?
