### 2022-08-16 AgrDataSci/climatrends#16 @dlebauer
installation of dev version fails
When following the installation in the README

Installing from CRAN works.

However, installing the development version fails - on both Windows and Ubuntu:

```
remotes::install_github("agrdatasci/climatrends")
```
I get the following errors:

> Downloading GitHub repo agrdatasci/climatrends@HEAD
> ✓  checking for file ‘/tmp/RtmpPZhckZ/remotes192662211476/AgrDataSci-climatrends-cddf697/DESCRIPTION’ (372ms)
> ─  preparing ‘climatrends’:
> ✓  checking DESCRIPTION meta-information ...
> ─  installing the package to process help pages
>          -----------------------------------
> ─  installing *source* package ‘climatrends’ ...
>    ** using staged installation
>    ** R
>    ** data
>    *** moving datasets to lazyload DB
>    ** inst
>    ** byte-compile and prepare package for lazy loading
>    Error in library("nasapower") : there is no package called ‘nasapower’
>    Error: unable to load R code in package ‘climatrends’
>    Execution halted
>    ERROR: lazy loading failed for package ‘climatrends’
> ─  removing ‘/tmp/RtmpyuDRgM/Rinst197f6af92b2/climatrends’
>          -----------------------------------
>    ERROR: package installation failed
> Error: Failed to install 'climatrends' from GitHub:
>   System command 'R' failed, exit status: 1, stdout + stderr (last 10 lines):
> E> *** moving datasets to lazyload DB
> E> ** inst
> E> ** byte-compile and prepare package for lazy loading
> E> Error in library("nasapower") : there is no package called ‘nasapower’
> E> Error: unable to load R code in package ‘climatrends’
> E> Execution halted
> E> ERROR: lazy loading failed for package ‘climatrends’
> E> * removing ‘/tmp/RtmpyuDRgM/Rinst197f6af92b2/climatrends’
> E>       -----------------------------------
> E> ERROR: package installation failed
