### 2022-08-16 #4405 @dlebauer
@KristinaRiemer trying, but currently failing to build https://github.com/AgrDataSci/climatrends/issues/16
