### 2022-08-16 #4405 @kauedesousa
I am working on updates but got stuck in other adm things. Better if you try the CRAN version which is more stable. Another issue is that NASAPower, which was used for the Scandinavia example, changed the API and don't allow calls for more than 1 year. The issue was already reported to the 'nasapower' developer, but there is not so much he can do https://github.com/ropensci/nasapower/issues/71. I am working in a solution from the climatrends side which probably will envolve a loop of lapply.
