### 2022-08-16 83c7c67
Update get_timeseries.R
move .nasapower() example to roxygen comments 
fixes #16
