### 2022-08-16 AgrDataSci/climatrends#18 @dlebauer
can not find function 'get_power'
I get an error when I run this code from the vignette:

``` r
library(climatrends)
    lonlat <- data.frame(lon = 129.19,
                         lat = 36.39)

    GDD(lonlat,
        day.one = "2019-04-01",
        last.day = "2019-10-01",
        degree.days = 150,
        return.as = "ndays")
#> Getting climate data from NASA POWER
#> Error in get_power(community = "ag", lonlat = c(129, 36, 132, 39), pars = c("T2M_MAX", : could not find function "get_power"
```
