+++ inst/paper/cbean.R
@@ -1,4 +1,5 @@
 library("climatrends")
+library("gosset")
 library("tidyverse")
 library("PlackettLuce")
 library("patchwork")
@@ -6,19 +7,19 @@ library("qvcalc")
 library("ggparty")
 
 
-load("inst/paper/cbean.rda")
+data("beans", package = "PlackettLuce")
 
 
 # compute the number of days required to accumulate gdd from 
 # planting date to maturity
 gdd <- GDD(modis, 
-           day.one = cbean$planting_date, 
+           day.one = beans$planting_date, 
            degree.days = 900, 
            return.as = "ndays")
 
 # add gdd to the cbean data
 # and take the average of gdd per season
-cbean %<>%  
+beans %<>%  
   mutate(gdd = gdd$gdd) %>% 
   group_by(season) %>% 
   mutate(gdds = as.integer(mean(gdd)))
@@ -26,20 +27,17 @@ cbean %<>%
 # compute the temperature indices from planting date to the number of days 
 # required to accumulate the gdd in each season
 temp <- temperature(modis, 
-                    day.one = cbean$planting_date, 
-                    span = cbean$gdds)
-
-rain <- rainfall(chirps, 
-                    day.one = cbean$planting_date, 
-                    span = cbean$gdds)
-
+                    day.one = beans$planting_date, 
+                    span = beans$gdds)
 
 # round values to 3 decimals, this will not change the main result, but will 
 # help in visualizing the tree 
 temp[1:ncol(temp)] <- lapply(temp[1:ncol(temp)], function(x) round(x, 2))
 
 # combine the indices with the main data
-cbean <- cbind(cbean, temp, rain)
+G <- rank_tricot(beans, 
+                 items = 
+                 input = )
 
 names(cbean)
 
