+++ vignettes/climatrends.bib
@@ -109,7 +109,6 @@ edition = {Training m},
 mendeley-groups = {gosset},
 publisher = {Food and Agriculture Organization of The United Nations},
 title = {{Irrigation water management: Irrigation water needs}},
-url = {http://www.fao.org/3/S2022E/s2022e00.htm},
 year = {1986}
 }
 
