### 2022-10-21 agrobioinfoservices/climatrends#16 @dlebauer
fixed in https://github.com/AgrDataSci/climatrends/commit/150114906f187953ec3f46a068549631b36170c2#diff-316f66d434fcc054db8f3cec81d677e23fe1edb948ce3665a3893f75a2b55d23R404
