### 2022-10-21 #4405 @dlebauer
@KristinaRiemer it installs now, though my review is still blocked by https://github.com/AgrDataSci/climatrends/issues/18
