+++ NAMESPACE
@@ -39,6 +39,7 @@ export(late_frost)
 export(rainfall)
 export(temperature)
 import(methods)
+importFrom(nasapower,get_power)
 importFrom(stats,cutree)
 importFrom(stats,dist)
 importFrom(stats,hclust)
