### 2022-12-10 agrobioinfoservices/climatrends#18 @kauedesousa
This issue is solved here https://github.com/AgrDataSci/climatrends/commit/83053e6c161b8396864bbeae7f86e22842612c45. The problem is that nasapower was not stated as a dependency for climatrends. Fixed.
