+++ inst/paper/paper.md
@@ -18,19 +18,17 @@ authors:
     affiliation: 2
   - name: Magne Neby
     orcid: 0000-0003-2098-7650
-    affiliation: "3"
+    affiliation: "1"
   - name: Svein Ø. Solberg
     orcid: 0000-0002-4491-4483
     affiliation: 1
 affiliations:
   - name: Department of Agricultural Sciences, Inland Norway University of Applied Sciences, Hamar, Norway
     index: 1
-  - name: Systems Transformation Group, Bioversity International, Montpellier, France
+  - name: Digital Inclusion, Bioversity International, Montpellier, France
     index: 2
-  - name: Department of Forest and Field Sciences, Inland Norway University of Applied Sciences, 2480 Koppang, Norway
-    index: 3
 citation_author: de Sousa et. al.
-year: 2022
+year: 2023
 bibliography: paper.bib
 output: rticles::joss_article
 journal: JOSS
@@ -44,7 +42,7 @@ journal: JOSS
 
 Abiotic factors play an important role in most ecological and crop systems that depend on certain levels of temperature, light and precipitation to initiate important physiological events [@PlantEcology]. Understanding how these factors drive the physiological processes is a key approach to provide recommendations for adaptation and biodiversity conservation in applied ecology studies. The package `climatrends` aims to provide the methods in R [@RCoreTeam] to compute precipitation and temperature indices that serve as input for climate and crop models [@vanEtten2019; @Kehel2016], trends in climate change [@Aguilar2005; @deSousa2018] and applied ecology [@Prentice1992; @YLiu2018].
 
-# Implementation
+# Statement of need
 
 Six main functions are provided (Table 1), with a default method for numeric 'vector' and additional methods implemented via the package `methods` [@RCoreTeam] for classes 'matrix' (or array), 'data.frame', and 'sf' (of geometry POINT or POLYGON) [@sf]. The last two methods are designed to fetch data from cloud sources, currently from the packages `nasapower` [@Sparks2018] and `chirps` [@chirps]. 
 
@@ -60,8 +58,6 @@ Table 1: Main functions available in climatrends.
 | temperature() | Temperature indices|
 
 These functions started as a set of scripts to compute indices from on-farm testing sites following a citizen science approach [@vanEtten2019]. Aiming to capture the environmental variation across different sites, which can differ as each on-farm trial generally have a different starting day and duration, the arguments `day.one` and `span` are vectorised and may be used to indicate the starting date for each data-point and the duration of the timespan to be considered for the computation of the indices. For time series analysis, fixed periods can be adjusted with the argument `last.day` linked to the argument `day.one`. 
-
-## Temperature and precipitation indices
 
 The package `climatrends` computes 12 temperature indices and 10 precipitation indices that were suggested by previous research on climatology and crop science [@Kehel2016; @Aguilar2005]. The indices computed by the functions `temperature()` and `rainfall()` are described in Table 2.
 
@@ -92,102 +88,26 @@ Table 2: Temperature and precipitation indices available in climatrends.
 |Rtotal | Total precipitation in wet days, rain >= 1 mm | mm |
 |SDII | Simple daily intensity index | mm/days |
 
-## Growing degree-days
-
-Growing degree-days (gdd) is an heuristic tool in phenology that measures heat accumulation and is used to predict plant and animal development rates [@Prentice1992]. Growing degree-days are calculated by taking the integral of warmth above a base temperature ($T_{0}$). The function `GDD()` applies by default the following equation.
-
-Equation [1]
-
-$$GDD = \frac{T_{max} + T_{min}}{2} - T_{0}$$
-
[…]
