+++ inst/paper/paper.bib
@@ -80,6 +80,23 @@ volume = {116},
 year = {2019}
 }
 
+@article{vanEtten2019tricot,
+author = {{{van Etten}}, Jacob and Beza, Eskender and Calderer, Llu{\'{i}}s and {{van Duijvendijk}}, Kees and Fadda, Carlo and Fantahun, Basazen and Kidane, Yosef Gebrehawaryat and {{van de Gevel}}, Jeske and Gupta, Arnab and Mengistu, Dejene Kassahun and Kiambi, Dan and Mathur, Prem Narain and Mercado, Leida and Mittra, Sarika and Mollel, Margaret J. and Rosas, Juan Carlos and Steinke, Jonathan and Suchini, Jose Gabriel and Zimmerer, Karl S.},
+doi = {10.1017/S0014479716000739},
+issn = {0014-4797},
+journal = {Experimental Agriculture},
+mendeley-groups = {gosset,phd-thesis,PPB},
+month = {jun},
+number = {S1},
+pages = {275--296},
+publisher = {Cambridge University Press},
+title = {{First experiences with a novel farmer citizen science approach: crowdsourcing participatory variety selection through on-farm triadic comparisons of technologies (tricot)}},
+url = {https://www.cambridge.org/core/product/identifier/S0014479716000739/type/journal_article},
+volume = {55},
+year = {2016}
+}
+
+
 
 @article{Wan2015,
 author = {Wan, Z. and Hook, S. and Hulley, G.},
