### 2023-01-04 agrobioinfoservices/climatrends#14 @kauedesousa
This is solved here https://github.com/AgrDataSci/climatrends/commit/74f09b5d693335e1a5cd4d01a0e9ff6c8d5d142e
