### 2023-01-04 agrobioinfoservices/climatrends#9 @kauedesousa
Solved here https://github.com/AgrDataSci/climatrends/commit/2d6528346287485fca1aea9f89ac01c71070c6e6 thanks
