### 2023-01-05 agrobioinfoservices/climatrends#13 @kauedesousa
NASAPOWER changed the API some months ago so they don't support calls for more than 366 days (a bit silly). This change affected this example and I've already reported it to the `nasapower` developers here https://github.com/ropensci/nasapower/issues/71. Since this problem is not originated in `climatrends` and not in `nasapower` either, and the solutions that I considered were not optimal for the package, I simply propose to wrap this in a loop, like the example below.

````
library("climatrends")
library("sf")
library("nasapower")
# create a polygon within the coordinates 7, 17, 59, 63
e <- matrix(c(7, 59, 17, 59, 17, 63,
              7, 63, 7, 59),
            nrow = 5, ncol = 2, byrow = TRUE)

e <- st_polygon(list(e))
# sample 100 points in the hexagonal type
p <- st_sample(e, 100, type = 'hexagonal')
p <- st_as_sf(p, crs = 4326)

dayone <- paste0(2000:2019, "-01-01")
dayone <- as.Date(dayone, origin = "1970-01-01")

temp <- data.frame()

for (d in seq_along(dayone)) {
  temp_d <- temperature(p,
                        day.one = dayone[d],
                        last.day = dayone[d] + 365,
                        timeseries = TRUE,
                        intervals = 365)

  temp <- rbind(temp, temp_d)

}

temp
````

But I think that this takes a lot of time to process, and to address the other issue that you rose here https://github.com/AgrDataSci/climatrends/issues/7 I think that we can put this example in one vignette and remove it from the paper.
