+++ inst/paper/paper.bib
@@ -7,6 +7,66 @@ url = {https://www.r-project.org/},
 year = {2020}
 }
 
+@article{Powers2019,
+annote = {doi: 10.1002/eap.1822},
+author = {Powers, Stephen M and Hampton, Stephanie E},
+doi = {10.1002/eap.1822},
+issn = {1051-0761},
+journal = {Ecological Applications},
+month = {jan},
+number = {1},
+pages = {e01822},
+publisher = {John Wiley & Sons, Ltd},
+title = {{Open science, reproducibility, and transparency in ecology}},
+url = {https://doi.org/10.1002/eap.1822},
+volume = {29},
+year = {2019}
+}
+
+@article{Stevens2017,
+author = {Stevens, Jeffrey R.},
+doi = {10.3389/fpsyg.2017.00862},
+issn = {1664-1078},
+journal = {Frontiers in Psychology},
+month = {may},
+pages = {862},
+publisher = {Frontiers Media S.A.},
+title = {{Replicability and Reproducibility in Comparative Psychology}},
+url = {http://journal.frontiersin.org/article/10.3389/fpsyg.2017.00862/full},
+volume = {8},
+year = {2017}
+}
+
+@Manual{agroclim,
+    title = {{agroclim: Climatic Indices for Agriculture}},
+    author = {Roberto Serrano-Notivoli},
+    year = {2022},
+    note = {R package version 0.3.0},
+    url = {https://CRAN.R-project.org/package=agroclim},
+  }
+
+@Manual{cropgrowdays,
+    title = {{cropgrowdays: Crop Growing Degree Days and Agrometeorological Calculations}},
+    author = {Peter Baker and Miranda Mortlock},
+    year = {2021},
+    note = {R package version 0.1.1},
+    url = {https://CRAN.R-project.org/package=cropgrowdays},
+  }
+
+@article{Munafo2017,
+author = {Munaf{\`{o}}, Marcus R and Nosek, Brian A and Bishop, Dorothy V M and Button, Katherine S and Chambers, Christopher D and {Percie du Sert}, Nathalie and Simonsohn, Uri and Wagenmakers, Eric-Jan and Ware, Jennifer J and Ioannidis, John P A},
+doi = {10.1038/s41562-016-0021},
+issn = {2397-3374},
+journal = {Nature Human Behaviour},
+number = {1},
+pages = {21},
+title = {{A manifesto for reproducible science}},
+url = {https://doi.org/10.1038/s41562-016-0021},
+volume = {1},
+year = {2017}
+}
+
+
 @article{chirps,
 author = {{{de Sousa}}, Kau{\^{e}} and Sparks, Adam and Ashmall, William and {{van Etten}}, Jacob and Solberg, Svein},
 doi = {10.21105/joss.02419},
@@ -278,6 +338,25 @@ mendeley-groups = {tricot-genomic},
 title = {{Modelling rankings in R: the PlackettLuce package}},
 url = {https://doi.org/10.1007/s00180-020-00959-3},
 year = {2020}
+}
+
+@article{Plackett1975,
+author = {Plackett, R. L.},
+doi = {10.2307/2346567},
+journal = {Journal of the Royal Statistical Society. Series C (Applied Statistics)},
+number = {2},
+pages = {193--202},
+title = {{The Analysis of Permutations}},
+volume = {24},
+year = {1975}
+}
+
+@book{Luce1959,
+author = {Luce, R. Duncan},
+pages = {153},
+publisher = {Courier Corporation},
+title = {{Individual Choice Behavior}},
+year = {1959}
 }
 
 @article{Wan2015,
