+++ inst/paper/paper.md
@@ -4,7 +4,6 @@ tags:
 - climate data
 - climatology
 - earth science
-- evapotranspiration
 - precipitation data
 - R
 - reproducibility
@@ -44,12 +43,14 @@ Abiotic factors play an important role in most ecological and crop systems that 
 
 # Statement of need
 
-Six main functions are provided (Table 1), with a default method for numeric 'vector' and additional methods implemented via the package `methods` [@RCoreTeam] for classes 'matrix' (or array), 'data.frame', and 'sf' (of geometry POINT or POLYGON) [@sf]. The last two methods are designed to fetch data from cloud sources, currently from the packages `nasapower` [@Sparks2018] and `chirps` [@chirps]. 
+Reproducibility, the ability to repeat the analysis, and replicability, the ability to repeat an experiment [@Stevens2017], are key to perform collaborative scientific research [@Powers2019; @Munafo2017]. It allows scientists to re-perform analysis after a long hiatus and the peers to validate analysis and get new insights using original or new data. This is still a gap in most of the studies in agriculture and ecology. To address this specific issue in reproducibility and replicability in utilizing climate data in agriculture and ecology we developed `climatrends`. The package contains eight main functions (Table 1), with a default method for numeric 'vector' and additional methods implemented via the package `methods` [@RCoreTeam] for classes 'matrix' (or array), 'data.frame', and 'sf' (of geometry POINT or POLYGON) [@sf]. The last two methods are designed to fetch data using API Client, currently from the packages `nasapower` [@Sparks2018] and `chirps` [@chirps]. 
 
 Table 1: Main functions available in climatrends.
 
 | **Function**  | **Definition**  |
 | --------- |----------------------|
+| cumdrought() | Returns a vector with the cumulative sum of the maximum length of dry spell (MLDS) |
+| cumrain() | Returns a vector with the cumulative sum of the maximum length of wet spell (MLWS) |
 | crop_sensitive() | Compute crop sensitive indices |
 | ETo() | Reference evapotranspiration using the Blaney-Criddle method|
 | GDD() | Compute growing degree-days|
@@ -57,45 +58,15 @@ Table 1: Main functions available in climatrends.
 | rainfall() | Precipitation indices|
 | temperature() | Temperature indices|
 
-These functions started as a set of scripts to compute indices from on-farm testing sites following a citizen science approach [@vanEtten2019]. Aiming to capture the environmental variation across different sites, which can differ as each on-farm trial generally have a different starting day and duration, the arguments `day.one` and `span` are vectorised and may be used to indicate the starting date for each data-point and the duration of the timespan to be considered for the computation of the indices. For time series analysis, fixed periods can be adjusted with the argument `last.day` linked to the argument `day.one`. 
[…]
