### 2023-01-05 agrobioinfoservices/climatrends#13 @kauedesousa
A new vignette was added here https://github.com/AgrDataSci/climatrends/commit/84bc35a13e7dde8af9802201aad7151abbaaea1d and the example was removed from the paper
