### 2023-01-05 agrobioinfoservices/climatrends#8 @kauedesousa
A comparison with other software is added to new paper version here https://github.com/AgrDataSci/climatrends/commit/39b2d0e952d7069e011af421eb0219f018a0c160
