### 2023-01-05 agrobioinfoservices/climatrends#7 @kauedesousa
A shorter version of the paper is added here https://github.com/AgrDataSci/climatrends/commit/39b2d0e952d7069e011af421eb0219f018a0c160
