### 2023-01-05 agrobioinfoservices/climatrends#6 @kauedesousa
Thanks! The comment is addressed here https://github.com/AgrDataSci/climatrends/issues/11
