+++ data-raw/DATASET.R
@@ -1,0 +1,5 @@
+## code to prepare `DATASET` dataset goes here
+
+usethis::use_data(commonbean, overwrite = TRUE, compress = "xz")
+
+dimnames(commonbean[[2]])
