+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: climatrends
 Type: Package
 Title: Climate Variability Indices for Ecological Modelling
-Version: 0.4-0003
+Version: 0.4-1
 Authors@R: c(person("Kauê", "de Sousa", 
         email = "desousa.kaue@gmail.com", role = c("aut", "cre"),
         comment = c(ORCID = "0000-0002-7571-7845")),
@@ -26,7 +26,6 @@ Suggests:
     ag5Tools,
     chirps,
     knitr,
-    nasapower,
     rmarkdown,
     sf,
     testthat
