### 2023-01-05 #4405 @kauedesousa
Dear all, thanks a lot for your comments and patience handling this submission. The last year was quite tough with many activities to handle and basically no time to write.

I addressed the comments and issues raised by @istfer and @dlebauer from which I am thankful. You can see them incorporated in climatrends v0.5 https://github.com/AgrDataSci/climatrends/commit/79b7c9f3fe08dbaa150d6cb05bcf9b98613d0111. I tested as CRAN and got one note on the size of the package which became larger to accommodate the bean data as suggested in one of the comments https://github.com/AgrDataSci/climatrends/issues/14. I will try to send it to CRAN as it is and see if they are convinced in publishing the package alongside the data. Otherwise I will try to find other solution.

The paper is significantly shorter than the previous version to address another issue https://github.com/AgrDataSci/climatrends/issues/7 but also includes the Statement of need as pointed out by the reviewers https://github.com/AgrDataSci/climatrends/issues/11

One of the examples in the paper was affected by a major change in the NASAPOWER API, and was pointed out by the reviewers https://github.com/AgrDataSci/climatrends/issues/13. This issue, however, goes beyond our control. As a solution, we removed the example from the paper and added it as a vignette `vignette("Climatology", "climatrends")`

Thanks again and please let me know if you need anything else from my side

Kauê
