### 2023-01-05 cfe75c3
enable-return-matrix-in-get_timeseries
