+++ NAMESPACE
@@ -15,7 +15,6 @@ S3method(crop_sensitive,sf)
 S3method(get_timeseries,array)
 S3method(get_timeseries,default)
 S3method(get_timeseries,matrix)
-S3method(get_timeseries,sf)
 S3method(late_frost,array)
 S3method(late_frost,data.frame)
 S3method(late_frost,default)
