+++ inst/paper/paper.md
@@ -43,7 +43,7 @@ Abiotic factors play an important role in most ecological and crop systems that 
 
 # Statement of need
 
-Reproducibility, the ability to repeat the analysis, and replicability, the ability to repeat an experiment [@Stevens2017], are key to perform collaborative scientific research [@Powers2019; @Munafo2017]. It allows scientists to re-perform analysis after a long hiatus and the peers to validate analysis and get new insights using original or new data. This is still a gap in most of the studies in agriculture and ecology. To address this specific issue in reproducibility and replicability in utilizing climate data in agriculture and ecology we developed `climatrends`. The package contains eight main functions (Table 1), with a default method for numeric 'vector' and additional methods implemented via the package `methods` [@RCoreTeam] for classes 'matrix' (or array), 'data.frame', and 'sf' (of geometry POINT or POLYGON) [@sf]. The last two methods are designed to fetch data using API Client, currently from the packages `nasapower` [@Sparks2018] and `chirps` [@chirps]. 
+Reproducibility, the ability to repeat the analysis, and replicability, the ability to repeat an experiment [@Stevens2017], are key to perform collaborative scientific research [@Powers2019; @Munafo2017]. It allows scientists to re-perform analysis after a long hiatus and the peers to validate analysis and get new insights using original or new data. This is still a gap in most of the studies in agriculture and ecology. `climatrends` addresses this specific issue. The package originates from a set of scripts to compute climate indices in our previous studies [@deSousa2018; @vanEtten2019]. Building up on the interest in expanding the analysis to other regions and to enable reproducible and replicable studies among different research groups across the CGIAR (https://www.cgiar.org) and partner institutions we developed `climatrends`. Most of the package functions take into account the heterogeneity of testing sites (locations), dates and seasons, a common characteristic of decentralized agricultural trials [@vanEtten2019]. Further development was made to enable time series analysis with fixed periods of time and locations [@deSousa2018]. The package `climatrends` computes temperature and precipitation indices that were suggested by previous studies on climatology and crop science [@Kehel2016; @Aguilar2005; @Challinor2016]. The indices are described in the package's documentation. `climatrends` was released on CRAN on February 2020 and, by that time, was the first package to provide methods for agroclimatic indices in R. Currently the package belongs to the CRAN Task View in Agriculture (https://cran.r-project.org/web/views/Agriculture.html) and is complementary to other new packages like `agroclim` [@agroclim] and `cropgrowdays` [@cropgrowdays].
 
 Table 1: Main functions available in climatrends.
 
@@ -58,11 +58,7 @@ Table 1: Main functions available in climatrends.
 | rainfall() | Precipitation indices|
 | temperature() | Temperature indices|
 
[…]
