### 2023-04-11 agrobioinfoservices/climatrends#11 @istfer
This PR adds the "statement of need" to the paper, but the checklist also requires it to be in the documentation, could you add it there too?
