### 2023-04-11 #4405 @istfer
Thank you @kauedesousa, most of my comments and suggestions have been addressed as well. Remaining two points are:

1. the JOSS checklist suggests that a similar statement of need section should also be in the package documentation. I saw that your PR only adds it to the paper, please let me know if I missed something. Otherwise, please add a similar section to the documentation as well.

2. The JOSS checklist also suggests that authors should describe how this software compares to other commonly-used packages. The only sentence that does something like this in the paper is as follows:

> "Currently the package belongs to the CRAN Task View in Agriculture (https://cran.r-project.org/web/views/Agriculture.html) and is complementary to other new packages like agroclim (Serrano-Notivoli, 2022) and cropgrowdays (Baker & Mortlock, 2021)."

I'm not sure if this suffices as the comparison to the state of the field. It would be nice to state at least in what ways it is complementary to the other packages. But I leave it to the @KristinaRiemer as the handling editor in the end, if she thinks this is good enough, I'm also fine by it.
