### 2023-04-11 #4405 @istfer
Hi @kauedesousa,

I thought adding a similar statement of need, like the one in the paper, somewhere [here](https://github.com/AgrDataSci/climatrends/blob/master/README.md) or [here](https://github.com/AgrDataSci/climatrends/blob/master/vignettes/Overview.Rmd) should be enough.
