+++ inst/paper/paper.bib
@@ -37,6 +37,14 @@ volume = {8},
 year = {2017}
 }
 
+@Manual{pollen,
+    title = {{pollen: Analysis of Aerobiological Data}},
+    author = {Jakub Nowosad},
+    year = {2019},
+    note = {R package version 0.71},
+    url = {https://CRAN.R-project.org/package=pollen},
+  }
+
 @Manual{agroclim,
     title = {{agroclim: Climatic Indices for Agriculture}},
     author = {Roberto Serrano-Notivoli},
@@ -410,3 +418,32 @@ year = {2008}
     note = {R package version 1.0.1},
     url = {https://CRAN.R-project.org/package=patchwork},
   }
+  
+  
+@article{Brown2023,
+  doi = {10.1016/j.softx.2022.101267},
+  url = {https://doi.org/10.1016/j.softx.2022.101267},
+  year = {2023},
+  month = feb,
+  publisher = {Elsevier {BV}},
+  volume = {21},
+  pages = {101267},
+  author = {David Brown and Kau{\^{e}} {de Sousa} and Jacob {van Etten}},
+  title = {{ag5Tools: An R package for downloading and extracting agrometeorological data from the {AgERA}5 database}},
+  journal = {{SoftwareX}}
+}
+
+
+@article{Hersbach2020,
+  doi = {10.1002/qj.3803},
+  url = {https://doi.org/10.1002/qj.3803},
+  year = {2020},
+  month = jun,
+  publisher = {Wiley},
+  volume = {146},
+  number = {730},
+  pages = {1999--2049},
+  author = {Hans Hersbach and Bill Bell and Paul Berrisford and Shoji Hirahara and Andr{\'{a}}s Hor{\'{a}}nyi and Joaqu{\'{\i}}n Mu{\~{n}}oz-Sabater and Julien Nicolas and Carole Peubey and Raluca Radu and Dinand Schepers and Adrian Simmons and Cornel Soci and Saleh Abdalla and Xavier Abellan and Gianpaolo Balsamo and Peter Bechtold and Gionata Biavati and Jean Bidlot and Massimo Bonavita and Giovanna Chiara and Per Dahlgren and Dick Dee and Michail Diamantakis and Rossana Dragani and Johannes Flemming and Richard Forbes and Manuel Fuentes and Alan Geer and Leo Haimberger and Sean Healy and Robin J. Hogan and El{\'{\i}}as H{\'{o}}lm and Marta Janiskov{\'{a}} and Sarah Keeley and Patrick Laloyaux and Philippe Lopez and Cristina Lupu and Gabor Radnoti and Patricia Rosnay and Iryna Rozum and Freja Vamborg and Sebastien Villaume and Jean-Noël Th{\'{e}}paut},
+  title = {The {ERA}5 global reanalysis},
+  journal = {Quarterly Journal of the Royal Meteorological Society}
+}
