+++ inst/paper/paper.md
@@ -39,7 +39,7 @@ Abiotic factors play an important role in most ecological and crop systems that 
 
 # Statement of need
 
-Reproducibility, the ability to repeat the analysis, and replicability, the ability to repeat an experiment [@Stevens2017], are key to perform collaborative scientific research [@Powers2019; @Munafo2017]. It allows scientists to re-perform analysis after a long hiatus and the peers to validate analysis and get new insights using original or new data. This is still a gap in most of the studies in agriculture and ecology. `climatrends` addresses this specific issue. The package originates from a set of scripts to compute climate indices in our previous studies [@deSousa2018; @vanEtten2019]. Building up on the interest in expanding the analysis to other regions and to enable reproducible and replicable studies among different research groups within the CGIAR (https://www.cgiar.org) and partner institutions we developed `climatrends`. Most of the package functions take into account the heterogeneity of testing sites (locations), dates and seasons, a common characteristic of decentralized agricultural trials [@vanEtten2019]. Further development was made to enable time series analysis with fixed periods of time and locations [@deSousa2018]. The package `climatrends` computes temperature, precipitation, crop growing and crop stress indices that were validated by previous studies on climatology and crop science [@Kehel2016; @Aguilar2005; @Challinor2016]. The indices are described in the package's documentation. `climatrends` was released on CRAN in February 2020 and was the first package to provide methods for agroclimatic indices in R, being able to deal with site heterogeneity and time series analysis. `climatrends` is also unique in its integration with API Client data like NASA POWER [@Sparks2018], CHIRPS [@Funk2015; @chirps] and AgERA5 [@Hersbach2020; @Brown2023]. Other packages like `pollen` [@pollen] and `cropgrowdays` [@cropgrowdays] are limited to growing degree days, whereas the package `agroclim` [@agroclim] provides temperature and precipitation indices but is limited to zoning areas. Currently `climatrends` belongs to the CRAN Task View in Agriculture (https://cran.r-project.org/web/views/Agriculture.html) and is complementary to `agroclim` and `cropgrowdays`.
[…]
