### 2023-04-11 #4405 @kauedesousa
Hi @istfer the statement of need is added to the documentation as you suggested https://github.com/AgrDataSci/climatrends/commit/386a78cf9e6149b1b7959bee22915490b5d7ca9c

I also expanded a bit the description of the package and comparison with other softwares here https://github.com/AgrDataSci/climatrends/commit/54d4a88e26f621248fa834192c731ed67073e9ab

Thanks for the support!
