### 2023-05-01 #4405 @KristinaRiemer
Thank you for your reviews @istfer & @dlebauer!

@kauedesousa, the first pre-publication step is to finalize the proof. Feel free to review the proof for any typos. I request that you make these final minor edits:

- Replace "perform" with "performing" (line 19)
- Replace "the peers" with "peers" (line 21)
- Replace "gdd" with "growing degree-days" (line 67)
- Add in-line text formatting using backticks to package and function names in the Fig. 1 figure caption

Once all changes have been made, please make a new tagged release of the repo and provide the version number here. If you could also archive the repo and provide the DOI. Most folks use Zenodo for this. Thanks!
