### 2023-05-01 #4405 @kauedesousa
Done. Solved with \texttt{}.
