### 2023-05-03 #4405 @kthyng
@kauedesousa Can you take a look at the error above? Sounds like you have a reference in your .bib twice maybe?
