+++ inst/paper/paper.bib
@@ -367,16 +367,6 @@ title = {{Individual Choice Behavior}},
 year = {1959}
 }
 
-@article{Wan2015,
-author = {Wan, Z. and Hook, S. and Hulley, G.},
-doi = {10.5067/MODIS/MYD11A1.006},
-journal = {NASA EOSDIS LP DAAC},
-publisher = {NASA EOSDIS LP DAAC},
-title = {{MYD11A1 MODIS/Aqua Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006 [Data set]}},
-url = {10.5067/MODIS/MYD11A1.006},
-year = {2015}
-}
-
 @article{deMedeiros2016,
 author = {{de Medeiros}, Gerson Araujo and Daniel, Luiz Antonio and Fengler, Felipe Hashimoto},
 doi = {10.1155/2016/8065985},
